Initial implementation
Introduces an internal API `NamedValidationAdapter`, instances of which can be
mixed into a Logstash plugin to ensure that the named validator exists. When
validating, the implementation from Logstash core is used whenever available,
and the provided backport is used only when necessary. This allows backports to
approximate upstream functionality, which in turn frees new validators
intrroduced in Logstash core from constraints.

Using this API, this PR also introduces  `FieldReferenceValidationAdapter`, an
adapter providing an approximation to the `:field_reference` named validator
proposed in [elastic/logstash#12459][].

Using this pattern, we will be free to begin introducing more validators to
Logstash core, absorbing some of the [bespoke validators][] introduced
in plugin projects.

[elastic/logstash#12459]: https://github.com/elastic/logstash/pull/12459
[bespoke validators]: https://github.com/logstash-plugins/logstash-input-elasticsearch/blob/v4.8.1/lib/logstash/inputs/elasticsearch.rb#L440-L477